+++ src/amun/gpu/mblas/vector_wrapper.h
@@ -27,7 +27,7 @@ public:
   }
 
   __device__
-  VectorWrapper(T *ptr, uint size)
+  VectorWrapper(T *ptr, unsigned size)
   {
     size_ = size;
     data_ = ptr;
@@ -35,7 +35,7 @@ public:
   }
 
   __device__ __host__
-  uint size() const
+  unsigned size() const
   {
     return size_;
   }
@@ -55,21 +55,21 @@ public:
   }
 
   __device__
-  const T &operator[](uint i) const
+  const T &operator[](unsigned i) const
   {
     assert(i < size());
     return data()[i];
   }
 
   __device__
-  T &operator[](uint i)
+  T &operator[](unsigned i)
   {
     assert(i < size());
     return data()[i];
   }
 
   __device__
-  VectorWrapper<T> Offset(size_t offset)
+  VectorWrapper<T> Offset(unsigned offset)
   {
     T &ele = (*this)[offset];
     VectorWrapper<T> ret(&ele, size_ - offset);
@@ -85,7 +85,7 @@ public:
   }
 
 protected:
-  uint size_;
+  unsigned size_;
 
   T *data_;
   const T *dataConst_;
